% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_association_testing.R
\name{adding_to_results}
\alias{adding_to_results}
\title{Simple function that adds to existing list names of lists within lists.}
\usage{
adding_to_results(a, b, c, d)
}
\arguments{
\item{a}{Name of list to append.}

\item{b}{RDS of original results.}

\item{c}{new results}

\item{d}{name of the trait}
}
\value{
an appended list.
}
\description{
Simple function that adds to existing list names of lists within lists.
}
