% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{age_of_onset_pheno_creator}
\alias{age_of_onset_pheno_creator}
\title{Creates age of onset specific versions of existing phenotypes}
\usage{
age_of_onset_pheno_creator(d, a, b, c, DOB_df)
}
\arguments{
\item{d}{Name of age_of_onset phenotype}

\item{a}{Phenotype data-frame}

\item{b}{Non-sex specific phenotype to be edited}

\item{c}{vector of male IDs}

\item{DOB_df}{date of birth data frame}
}
\value{
Age of onset phenotype as list.
}
\description{
Creates age of onset specific versions of existing phenotypes
}
\keyword{internal}
