% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_phenotyping.R
\name{AND}
\alias{AND}
\title{Combines dataframes according to Boolean logical AND.}
\usage{
AND(a, b, c, d)
}
\arguments{
\item{a}{Left hand side of the argument, i.e. a AND b}

\item{b}{Right hand side of the argument, i.e. a AND b}

\item{c}{lower limit of the gap between dates of the first code of a and b}

\item{d}{upper limit of the gap between dates of the first code of a and b}
}
\value{
A dataframe of filtered ID's based on date relationship and AND Boolean logic.
}
\description{
Combines dataframes according to Boolean logical AND.
}
