% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_creation.R
\name{calc_minus_source_ID}
\alias{calc_minus_source_ID}
\title{Minor function for concept creation}
\usage{
calc_minus_source_ID(x, included_ID)
}
\arguments{
\item{x}{numeric list}

\item{included_ID}{list of codes}
}
\value{
A dataframe containing of extracted data, each data item is a row with participant ID, code and source
}
\description{
Minor function for concept creation
}
