% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_creation.R
\name{clinical_code_lookup}
\alias{clinical_code_lookup}
\title{Creates clinical concepts for combination in composite phenotypes}
\usage{
clinical_code_lookup(x, z, health_data, code_list_folder)
}
\arguments{
\item{x}{PheWAS_ID for concept}

\item{z}{location of the file}

\item{health_data}{dataframe containing health_data}

\item{code_list_folder}{marker to detect if default concept data is used or alternative folder of concept code lists is provided}
}
\value{
A list each containing 2 lists one for WDA and one for all dates used later in the pipeline for clinical concepts
}
\description{
Creates clinical concepts for combination in composite phenotypes
}
\keyword{internal}
