% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_split.R
\name{combine_split}
\alias{combine_split}
\title{Combines PLINK results from the split_groups and split_analysis into a single file.}
\usage{
combine_split(
  split_plink_results_folder,
  results_RDS_file,
  group_name,
  analysis_name
)
}
\arguments{
\item{split_plink_results_folder}{Full path of the folder containing all of the PLINK results from the split_group, split_analysis option in 03a_PLINK_association_testing.R.}

\item{results_RDS_file}{Full path of the results of the analysis from the 03a_PLINK_association_testing.R script. Is an R object.}

\item{group_name}{Name of the group that was split that the results represent.}

\item{analysis_name}{Name for the analysis to be used in the naming of the output files, use the same as in 03a_PLINK_association_analysis.}
}
\value{
an updated results RDS file that includes the now combined group results.
}
\description{
Combines PLINK results from the split_groups and split_analysis into a single file.
}
