% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_phenotyping.R
\name{combining_codes}
\alias{combining_codes}
\title{Combines concepts for use in creating composite phenotypes}
\usage{
combining_codes(x, y, phenotypes)
}
\arguments{
\item{x}{Concepts to extract}

\item{y}{N_codes used to extract concepts}

\item{phenotypes}{all available phenotypes generated by the previous functions.}
}
\value{
A list of extracted concepts filtered for N_codes
}
\description{
Combines concepts for use in creating composite phenotypes
}
