% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_field_phenotype_generation.R
\name{combining_data_field}
\alias{combining_data_field}
\title{Makes combined data-field phenotypes from study data guided by PheWAS manifest}
\usage{
combining_data_field(
  a,
  b,
  c,
  d,
  e,
  f,
  PheWAS_manifest,
  current_data_fields,
  data_field_variables
)
}
\arguments{
\item{a}{PheWAS_ID}

\item{b}{selecting whether first_last_max_min_value}

\item{c}{existence of limits or not}

\item{d}{lower_limit quant value}

\item{e}{upper_limit quant value}

\item{f}{data coding for age}

\item{PheWAS_manifest}{PheWAS_manifest}

\item{current_data_fields}{vector of current data fields available generated from previous code}

\item{data_field_variables}{dataframe of extracted data fields}
}
\value{
A dataframe containing combined data-field phenotype
}
\description{
Makes combined data-field phenotypes from study data guided by PheWAS manifest
}
\keyword{internal}
