% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_phenotyping.R
\name{composite_phenotyping}
\alias{composite_phenotyping}
\title{Makes composite phenotypes from study data guided by composite phenotype map.}
\usage{
composite_phenotyping(
  composite_phenotype_map_overide,
  phenotype_folder,
  phenotype_files,
  N_iterations,
  phenotype_save_file,
  control_populations,
  update_list,
  control_pop_save_file = NULL
)
}
\arguments{
\item{composite_phenotype_map_overide}{Full path of the composite_phenotype_map file. Provided with the package.}

\item{phenotype_folder}{Full path of the folder containing the phenotype data created in previous steps.}

\item{phenotype_files}{Comma separated full file paths of phenotype data created in previous steps.}

\item{N_iterations}{Number of iterations curated phenotype script is to run through. Default: 5}

\item{phenotype_save_file}{Full path for the save file for the generated composite phenotypes RDS.}

\item{control_populations}{Full path of the "control_populations" file containing columns of IDs, column names are used to create lists of IDs. Saved as a list called control_populations.RDS in the folder inputted in the phenotype_folder flag. The lists are used to define some of the composite phenotype control populations as directed by the composite_phenotype_map file. The unedited version of the composite_phenotype_map file uses two populations, all_pop and primary_care_pop, which represent all IDs in the sample and all IDs with available primary care data. To create composite phenotypes, the names of these list must match the composite_phenotype_map file. The 02_data_preparation.R script creates a control_populations file that is used by default.}

\item{update_list}{Option to run this script as an update to an existing composite_phenotype list object. If specified, this script will use phenotype_save_file to load the existing list and then save over that file upon completion.}

\item{control_pop_save_file}{Full path for the save file for the generated composite control populations RDS.}
}
\value{
An RDS file containing lists of data-frames each representing a single composite phenotype.
}
\description{
Makes composite phenotypes from study data guided by composite phenotype map.
}
