% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_creation.R
\name{concept_source_data_extraction}
\alias{concept_source_data_extraction}
\title{Extracts codes for concepts into the various versions required for other functions}
\usage{
concept_source_data_extraction(a, code_list, health_data)
}
\arguments{
\item{a}{source of data}

\item{code_list}{list of codes}

\item{health_data}{dataframe containing health_data}
}
\value{
A dataframe containing of extracted data, each data item is a row with participant ID, code and source
}
\description{
Extracts codes for concepts into the various versions required for other functions
}
\keyword{internal}
