% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_creation.R
\name{creating_concepts}
\alias{creating_concepts}
\title{Creates prescription concepts for combination in composite phenotypes}
\usage{
creating_concepts(
  health_data,
  GPP,
  concept_save_file,
  all_dates_save_file,
  PheWAS_manifest_overide,
  code_list_folder_override
)
}
\arguments{
\item{health_data}{Full file path of the min_data file. Selecting will generate clinical concepts.}

\item{GPP}{Full file path of the GP prescription data. Selecting will generate prescription concepts.}

\item{concept_save_file}{Full file path for the save file for the generated concepts RDS used for concept creation.}

\item{all_dates_save_file}{Full file path for the save file for the generated all_dates.RDS used for per-event combinations of concepts.}

\item{PheWAS_manifest_overide}{Full file path of the alternative PheWAS_manifest file.}

\item{code_list_folder_override}{Full file for the folder containing code lists only use if not using default stored in R package.}
}
\value{
Two RDS objects for concepts one containing all_dates and one containing summarised N codes and earliest date.
}
\description{
Creates prescription concepts for combination in composite phenotypes
}
