% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_phenotyping.R
\name{curated_phenotype_creation}
\alias{curated_phenotype_creation}
\title{Makes composite phenotypes guided by composite phenotype map.}
\usage{
curated_phenotype_creation(a, x, phenotypes, curated_pheno_list)
}
\arguments{
\item{a}{PheWAS_IDs}

\item{x}{instructions passed from the composite phenotype map}

\item{phenotypes}{all available phenotypes generated by the previous functions.}

\item{curated_pheno_list}{composite phenotypes that have been made starts with none}
}
\value{
A list of data frames of composite phenotypes
}
\description{
Makes composite phenotypes guided by composite phenotype map.
}
