% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_field_phenotype_generation.R
\name{data_field_extraction}
\alias{data_field_extraction}
\title{Extracts data-field phenotypes from study data guided by PheWAS manifest}
\usage{
data_field_extraction(a, b, c, d, e, f, g, h, i, j, k, l, m, n, min_data)
}
\arguments{
\item{a}{PheWAS_ID}

\item{b}{QC_flag_ID}

\item{c}{search_id}

\item{d}{type of analysis}

\item{e}{data coding for case}

\item{f}{data coding for controls}

\item{g}{existence of limits or not}

\item{h}{lower_limit quant value}

\item{i}{upper_limit quant value}

\item{j}{data coding for date}

\item{k}{data coding for age}

\item{l}{selecting whether first_last_max_min_value}

\item{m}{selecting whether to exclude_case_control_crossover}

\item{n}{data coding for QC vales}

\item{min_data}{dataframe of min_data}
}
\value{
A dataframe containing data field phenotype
}
\description{
Extracts data-field phenotypes from study data guided by PheWAS manifest
}
\keyword{internal}
