% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_field_phenotype_generation.R
\name{data_field_phenotyping}
\alias{data_field_phenotyping}
\title{Makes data-field phenotypes from study data guided by PheWAS manifest}
\usage{
data_field_phenotyping(
  min_data,
  phenotype_save_file,
  N_cores,
  PheWAS_manifest_overide,
  append_file
)
}
\arguments{
\item{min_data}{Full path of the tab-separated file generated by the minimum_data.R script.}

\item{phenotype_save_file}{Full path for the save file for the generated concepts RDS to be used for phenotype creation.}

\item{N_cores}{Number of cores requested if parallel computing is desired. Defaults to single core computing.}

\item{PheWAS_manifest_overide}{Full path of an alternative to the in built file.}

\item{append_file}{Full path of an existing R object containing Data-Field phenotype information to which additional phenotypes derived by this script will be appended.}
}
\value{
An RDS file containing lists of data-frames each representing a single data_field phenotype containing combined data-field phenotypes
}
\description{
Makes data-field phenotypes from study data guided by PheWAS manifest
}
