% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{data_preparation_R}
\alias{data_preparation_R}
\title{Manipulates data from UK Biobank and creates files that are used in the later stages of phenotype development.}
\usage{
data_preparation_R(
  min_data,
  GPC,
  GPP,
  hesin_diag,
  HESIN,
  hesin_oper,
  death_cause,
  death,
  exclusions,
  king_coef,
  save_location
)
}
\arguments{
\item{min_data}{Full path of the file generated by the previous step (01_minimum_data.R).}

\item{GPC}{Full path of the primary care clinical data from UK Biobank.}

\item{GPP}{Full path of the primary care prescription data from UK Biobank.}

\item{hesin_diag}{Full file path of hesin_diag file from UK-Biobank.}

\item{HESIN}{Full file path of HESIN file from UK-Biobank.}

\item{hesin_oper}{Full file path of the hesin_oper file from UK-Biobank.}

\item{death_cause}{Full file path of the death_cause file from UK-Biobank.}

\item{death}{Full file path of the death file from UK-Biobank contains date of death info.}

\item{exclusions}{Full path of the file containing individuals to be excluded from the analysis. Defaults behaviour is to retain all individuals.}

\item{king_coef}{Full file path of related data file containing King coefficient scores for related ID pairs from UK-Biobank.}

\item{save_location}{Full file path for the common folder to save created files.}
}
\value{
Returns a series of files dependent on the inputs.
}
\description{
Manipulates data from UK Biobank and creates files that are used in the later stages of phenotype development.
}
