% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables_graphs.R
\name{Deep_PheWAS_graphs_tables}
\alias{Deep_PheWAS_graphs_tables}
\title{Makes graphs from converted RDS files.}
\usage{
Deep_PheWAS_graphs_tables(
  x,
  y,
  save_folder,
  PheWAS_ID_filter,
  plink_results,
  save_raw_plink,
  analysis_name,
  SNP_list,
  group_name_filter,
  no_save_table_all_results,
  SNP_filter,
  MAC,
  PheWAS_manifest,
  updated_manifest,
  sex_split,
  save_table_per_group_name,
  save_table_per_snp,
  R_association_graph,
  graph_choice,
  R_association_results,
  per_group_name_graph,
  per_snp_graph,
  max_pheno_tests,
  FDR_figure,
  max_FDR_graph,
  MAC_figure,
  MAC_cases_N,
  MAC_control_N,
  PheWAS_label_filter,
  max_overlap,
  graph_type,
  label_size
)
}
\arguments{
\item{x}{Group name}

\item{y}{Result file}

\item{save_folder}{Location of folder to save results}

\item{PheWAS_ID_filter}{List of PheWAS_IDs to filter tables and graph production.}

\item{plink_results}{Logical value if results are from Plink.}

\item{save_raw_plink}{Logical value indicating if wanting to save the raw plink results.}

\item{analysis_name}{name of analysis used in non-GRS data only.}

\item{SNP_list}{The list of SNPs used in extract_SNPs script (if used).}

\item{group_name_filter}{Name of group(s) to filter results by.}

\item{no_save_table_all_results}{logical if True then do not save all results.}

\item{SNP_filter}{List of SNPs to filter the graphs and tables production to.}

\item{MAC}{Minor allele count filter.}

\item{PheWAS_manifest}{PheWAS manifest data frame.}

\item{updated_manifest}{Altered PheWAS manifest to create all possible phenotypes.}

\item{sex_split}{Logical, if True three graphs are produced per input two sex stratified graphs and one graph with the original phenotype.}

\item{save_table_per_group_name}{Logical if True saves tables per group.}

\item{save_table_per_snp}{Logical if True saves tables per SNP.}

\item{R_association_graph}{Logical if True indicates use different graph input and data is from R-association rather than Plink.}

\item{graph_choice}{Value that indicates whether to produce one or two graphs and which ones.}

\item{R_association_results}{Logical if True indicates data is from R-association rather than Plink.}

\item{per_group_name_graph}{Logical if true produce graphs grouping by group.}

\item{per_snp_graph}{Logical if true produce graphs grouping by SNP}

\item{max_pheno_tests}{the maximum number of tests that are being run across grouping variable. Must be at => the number of associations.}

\item{FDR_figure}{value of significance for FDR.}

\item{max_FDR_graph}{value for lowest fdr value, converts 0 into this number.}

\item{MAC_figure}{Minor allele count filter numeric.}

\item{MAC_cases_N}{Minor allele count filter in cases}

\item{MAC_control_N}{Minor allele count filter in controls}

\item{PheWAS_label_filter}{list of PheWAS_IDs to label within the all_pheno graph.}

\item{max_overlap}{maximum number of overlaps for labelled phenotypes in the all_pheno graphs}

\item{graph_type}{save format of the graphs any input readable from ggsave is accepted}

\item{label_size}{size of the text for labelled phenotypes.}
}
\value{
Saved tables and graphs, graphs created through calling additional functions.
}
\description{
Makes graphs from converted RDS files.
}
