% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concept_creation.R
\name{drug_code_lookup}
\alias{drug_code_lookup}
\title{Creates prescription concepts for combination in composite phenotypes}
\usage{
drug_code_lookup(x, y, z, GP_P, V2_drugs, code_list_folder)
}
\arguments{
\item{x}{PheWAS_ID for concept}

\item{y}{Read_V2 drug list}

\item{z}{Drug key search terms}

\item{GP_P}{dataframe containing GP prescription data}

\item{V2_drugs}{dataframe containing information on drugs coded using Read_V2 codes.}

\item{code_list_folder}{marker to detect if default concept data is used or alternative folder of concept code lists is provided}
}
\value{
A list each containing 2 lists one for WDA and one for all dates used later in the pipeline for prescription concepts
}
\description{
Creates prescription concepts for combination in composite phenotypes
}
