% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phecode_generating.R
\name{exclusion_generator}
\alias{exclusion_generator}
\title{Creates exclusion ranges which are then used for composite phenotypes}
\usage{
exclusion_generator(a, b, c, d, e, f, g, h)
}
\arguments{
\item{a}{Name of range exclusion file}

\item{b}{1st lower phecode range}

\item{c}{1st upper phecode range}

\item{d}{2nd lower phecode range}

\item{e}{2nd upper phecode range}

\item{f}{3rd lower phecode range}

\item{g}{3rd upper phecode range}

\item{h}{phecodes_mapped data frame}
}
\value{
A dataframe containing ids of participants to be excluded as controls for given set of phenotypes
}
\description{
Creates exclusion ranges which are then used for composite phenotypes
}
