% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables_graphs.R
\name{fdr_calc}
\alias{fdr_calc}
\title{Converts P values into false discovery rate.}
\usage{
fdr_calc(x, max_pheno_tests)
}
\arguments{
\item{x}{Group name}

\item{max_pheno_tests}{the maximum number of tests that are being run across grouping variable. Must be at => the number of associations.}
}
\value{
data frame with new fdr column.
}
\description{
Converts P values into false discovery rate.
}
