% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula_phenotype_generation.R
\name{formula_phenotypes}
\alias{formula_phenotypes}
\title{Creates formula phenotypes.}
\usage{
formula_phenotypes(
  min_data,
  data_field_phenotypes,
  sex_info,
  PheWAS_manifest_overide,
  phenotype_save_file
)
}
\arguments{
\item{min_data}{Full path of the file generated by the previous step (01_minimum_data.R).}

\item{data_field_phenotypes}{Full path of the data_field.RDS file produced by 03b_data_field_phenotypes.R, currently only used for eGFR phenotype.}

\item{sex_info}{Full file path of the combined_sex file a file containing sex information (0=female, 1=male). Used only for eGFR field_ID phenotype.}

\item{PheWAS_manifest_overide}{Full file path of the alternative PheWAS_manifest file.}

\item{phenotype_save_file}{Full path of the save file for the generated concepts RDS to be used for phenotype creation.}
}
\value{
A R object containing formula phenotypes.
}
\description{
Creates formula phenotypes.
}
