% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phecode_generating.R
\name{generating_phecodes}
\alias{generating_phecodes}
\title{Creates phecode phenotypes and saves as a list alongside range_exclusions}
\usage{
generating_phecodes(
  health_data,
  sex_info,
  no_range_ID,
  no_phecodes,
  control_exclusions,
  N_cores,
  phecode_save_file,
  range_ID_save_file,
  ICD10,
  ICD9,
  PheWAS_manifest_overide
)
}
\arguments{
\item{health_data}{Full path of the file produced by the previous step (phenotype_data.R).}

\item{sex_info}{Full file path of the combined_sex file a file containing participant ID and sex information (0=female, 1=male).}

\item{no_range_ID}{Choose whether to save lists of participant identifiers to be excluded from control definitions (FALSE) or not (TRUE). Defaults to FALSE.}

\item{no_phecodes}{Choose whether to save the Phecode-generated phenotype data (FALSE) or not (TRUE). Defaults to FALSE.}

\item{control_exclusions}{Full file path of the optional control exclusions file.}

\item{N_cores}{Number of cores requested if parallel computing is desired. Defaults to single core computing.}

\item{phecode_save_file}{Full path of the save file for the Phecode phenotypes R data object.}

\item{range_ID_save_file}{Full path of the folder used to store the control exclusions R data object.}

\item{ICD10}{Comma-separated string representing the column names used in health_data for ICD-10 values. If there are no ICD-10 values, use NA or any text not used as a source in health data. Defaults to "ICD10".}

\item{ICD9}{Comma-separated string representing the column names used in health_data for ICD-9 values. If there are no ICD-9 values, use NA or any text not used as a source in health data. Defaults to "ICD9".}

\item{PheWAS_manifest_overide}{Full file path of the alternative PheWAS_manifest file.}
}
\value{
A rds file containing phecode phenotypes.
}
\description{
Creates phecode phenotypes and saves as a list alongside range_exclusions
}
