% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables_graphs.R
\name{graphs_tables_DeepPheWAS}
\alias{graphs_tables_DeepPheWAS}
\title{Makes graphs and tables from RDS results files.}
\usage{
graphs_tables_DeepPheWAS(
  results_file,
  analysis_name,
  plink_results,
  SNP_list,
  R_association_results,
  save_folder,
  group_filter,
  PheWAS_ID_filter,
  PheWAS_manifest_overide,
  max_pheno,
  sig_FDR,
  no_save_table_all_results,
  no_graph_all,
  no_graph_sig,
  max_FDR_graph,
  sex_split,
  SNP_filter,
  group_name_filter,
  save_raw_plink,
  MAC,
  MAC_case,
  MAC_control,
  per_group_name_graph,
  per_snp_graph,
  save_table_per_group_name,
  save_table_per_snp,
  R_association_graph,
  PheWAS_ID_label_filter,
  max_overlap_labels,
  graph_file_save,
  label_text_size,
  order_groups_alphabetically,
  order_phenotypes_alphabetically,
  save_all_graphs
)
}
\arguments{
\item{results_file}{Full file path of the results file RDS R list object.}

\item{analysis_name}{Name for the analysis.}

\item{plink_results}{Select if results are from 03a_PLINK_association_testing.R}

\item{SNP_list}{Full path of the SNP_list file used in 02_extracting_snps.R.}

\item{R_association_results}{Select if results are from 03b_R_association_testing.R.}

\item{save_folder}{File path of the folder to which the output will be saved.}

\item{group_filter}{Comma-separated text input, used to filter the group to which the table and graph functions are applied. Inputted groups are the ones that are retained for analysis, group here refers to the grouping variable used to subset the analysis classically ancestry.}

\item{PheWAS_ID_filter}{Full file path to a plain txt file containing single column NO header containing full PheWAS_ID of phenotypes that will be included Use if wanting to apply the table and/or graphing functions a subset of phenotypes.}

\item{PheWAS_manifest_overide}{Full file path of the alternative PheWAS_manifest file.}

\item{max_pheno}{Manual override for inputting maximum phenotypes analysed. Used for calculating FDR. The default used the largest number of associations in the results_file of any grouping.}

\item{sig_FDR}{Value of FDR for which associations are reported as significant. Will alter the line of significance in both graph types and the reported phenotypes in the significant_pheno graphs. Default: 0.01.}

\item{no_save_table_all_results}{Specify whether a table containing all the results should be saved (FALSE) or not (TRUE). Default is FALSE.}

\item{no_graph_all}{Specify whether a figure of all the results should be produced (FALSE) or not (TRUE). Default is FALSE.}

\item{no_graph_sig}{Specify whether a figure of the significant results should be produced (FALSE) or not (TRUE). Default is FALSE.}

\item{max_FDR_graph}{Maximum value for the y-axis to be used when the FDR for a particular association is exceptionally small. Default is 300.}

\item{sex_split}{Specify whether the figure should be stratified by sex (TRUE) or not (FALSE). If TRUE, this will produce separate figures for males, females and combined. Default is FALSE. Does not create split tables.}

\item{SNP_filter}{Full path of a file containing SNP IDs, used to filter the output to a given subset of SNPs. Use if wanting to apply the table and/or graphing functions a subset of results. Only works for plink_results.}

\item{group_name_filter}{Full path of the file containing a list of group_names that correspond with those defined in SNP_list. This can be used if the table and/or graphing functions should be applied to a subset of SNPs, such as only those in a given credible set. This option can only be used with association results generated by 03a_PLINK_association_testing.R.}

\item{save_raw_plink}{Specify whether the unfiltered association results from the PLINK analysis should be saved (TRUE) or not (FALSE). Default is FALSE.}

\item{MAC}{MAC (minor allele count) filter applied to all associations, only applicable in results from03a_PLINK_association_testing.R Default is 20.}

\item{MAC_case}{Minor allele count among cases for filtering the association results. only applicable in results from 03a_PLINK_association_testing.R and only in binary phenotypes. Default is 5.}

\item{MAC_control}{Minor allele count among controls for filtering the association results. Only applicable in results from 03a_PLINK_association_testing.R and only in binary phenotypes. Default: 10.}

\item{per_group_name_graph}{Select if wanting to produce graphs per-group_name. This is used when looking to report the most significant finding across several SNPs for a single construct, potentially and gene or a sentinel SNP with a credible set. It is a column in the SNP_list file. Default is FALSE.}

\item{per_snp_graph}{Specify whether a figure is produced for every SNP provided (TRUE) or not (FALSE). Default is FALSE.}

\item{save_table_per_group_name}{Select if wanting to produce tables per-group_name. This is used when looking to report the most significant finding across several SNPs for a single construct, potentially and gene or a sentinel SNP with a credible set. It is a column in the SNP_list file. Default is FALSE.}

\item{save_table_per_snp}{Specify whether a results table should be generated for every SNP provided (TRUE) or not (FALSE). Default is FALSE. Will be saved in a created folder named /analysis_name_group_per_SNP_tables. Example if the group was groupA and analysis_name top_SNPs the folder would be /top_SNPs_groupA_per_SNP_tables.}

\item{R_association_graph}{Specify whether a figure of the results from the association analysis from 03b_R_association_testing.R should be produced (TRUE) or not (FALSE). Default is FALSE.}

\item{PheWAS_ID_label_filter}{Full file path to plain text file containing single headed column of PheWAS_IDs. Only these PheWAS_IDs will be labelled within any graphical output designed to be used primarily when trying to edit a single graph as the filter will apply to all graphs being created.}

\item{max_overlap_labels}{Number, represents maximum overlaps for labelling of phenotypes in the all_pheno graph, lowering the number has the effect of reducing the total number of phenotypes labelled. Default: 20}

\item{graph_file_save}{Allows user to specify the file format of the graphs, for example pdf or png. Default: png}

\item{label_text_size}{Number, represents the text size of the labelled phenotypes in all_pheno graph. Default: 2}

\item{order_groups_alphabetically}{Specify whether the groups in the graphs should be ordered by lowest FDR (FALSE) or alphabetically (TRUE).}

\item{order_phenotypes_alphabetically}{Specify whether the phenotypes within each group are ordered by the lowest FDR (FALSE) or alphabetically (TRUE).}

\item{save_all_graphs}{Specify whether to always save every graph with or without a significant result (TRUE) or to only save when at least one association is significant (FALSE).}
}
\value{
Tables and graphs of the association results.
}
\description{
Makes graphs and tables from RDS results files.
}
