% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{group_id_creater}
\alias{group_id_creater}
\title{Creates grouping ID's}
\usage{
group_id_creater(x, groupings_df)
}
\arguments{
\item{x}{Groups}

\item{groupings_df}{grouping data frame}
}
\value{
Grouping ID dataframes as lists
}
\description{
Creates grouping ID's
}
\keyword{internal}
