% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_association_testing.R
\name{GRS_association}
\alias{GRS_association}
\title{Runs R association testing per group per trait.}
\usage{
GRS_association(
  e,
  a,
  b,
  c,
  d,
  GRS_input,
  non_GRS_data,
  covariates,
  all_phenos,
  PheWAS_manifest,
  analysis_save_name,
  quant_min_cases,
  binary_min_cases,
  covariate_col_names
)
}
\arguments{
\item{e}{group identifier.}

\item{a}{Trait name.}

\item{b}{Full file path to location of the genetic data.}

\item{c}{Full file path to location of the phenotype data.}

\item{d}{column name within the genetic data to be extracted for association analysis.}

\item{GRS_input}{Data frame guide to location of GRS data, and phenotype data per trait/grouping combination.}

\item{non_GRS_data}{Data frame of non-GRS data, eid column, with each column representing a genetic variable.}

\item{covariates}{Data frame of covariates for analysis.}

\item{all_phenos}{Data frame containing all possible phenotypes including age coloumn covariates.}

\item{PheWAS_manifest}{Data frame of the PheWAS_manifest containing information on phenotypes.}

\item{analysis_save_name}{Core save name for the analysis.}

\item{quant_min_cases}{minimum number of quantitative cases.}

\item{binary_min_cases}{minimum number of binary cases.}

\item{covariate_col_names}{column names of covariates.}
}
\value{
Plink association results.
}
\description{
Runs R association testing per group per trait.
}
