% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_association_testing.R
\name{GRS_association_testing}
\alias{GRS_association_testing}
\title{Runs R association testing.}
\usage{
GRS_association_testing(
  a,
  b,
  c,
  d,
  phenotypes,
  PheWAS_manifest,
  GRS_cov,
  quant_min_cases,
  binary_min_cases,
  covariate_col_names,
  all_phenos
)
}
\arguments{
\item{a}{PheWAS_ID}

\item{b}{age column information whether generic age covariate or age at measurement extracted from the phenotype files.}

\item{c}{analysis option binary or quantitiative}

\item{d}{variable name}

\item{phenotypes}{data frame of all phenotypes for current grouping.}

\item{PheWAS_manifest}{Data frame of the PheWAS_manifest containing information on phenotypes.}

\item{GRS_cov}{covariates for association testing}

\item{quant_min_cases}{minimum cases for quantitative phenotypes}

\item{binary_min_cases}{minimum cases for binary phenotypes}

\item{covariate_col_names}{column names of covariates.}

\item{all_phenos}{Data frame containing all possible phenotypes including age coloumn covariates.}
}
\value{
A dataframe of association results per phenotype.
}
\description{
Runs R association testing.
}
