% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_association_testing.R
\name{inner_names}
\alias{inner_names}
\title{Simple function that extracts names of lists within lists.}
\usage{
inner_names(x, y)
}
\arguments{
\item{x}{Name of the first level of list.}

\item{y}{list where names are extracted.}
}
\value{
a list of names from each internal list
}
\description{
Simple function that extracts names of lists within lists.
}
