% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{IVNT_transformation}
\alias{IVNT_transformation}
\title{Applies transformations to phenotypes as directed by manifest file.}
\usage{
IVNT_transformation(x, y, PheWAS_manifest, age_of_onset_phenotypes)
}
\arguments{
\item{x}{Phenotypes as list object}

\item{y}{PheWAS_ID of phenotype}

\item{PheWAS_manifest}{the PheWAS manifest}

\item{age_of_onset_phenotypes}{either NULL or location of age_of_onset phenotypes file}
}
\value{
Phenotypes as list transformed as directed ready to convery into  a dataframe
}
\description{
Applies transformations to phenotypes as directed by manifest file.
}
