% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{join_cols_pheno}
\alias{join_cols_pheno}
\title{Allows main conversion to data frame of phenotypes}
\usage{
join_cols_pheno(
  x,
  IVNT,
  age_of_onset_phenotypes,
  PheWAS_manifest,
  age_of_onset_all,
  all_phenotypes
)
}
\arguments{
\item{x}{All phenotypes as list objects}

\item{IVNT}{logical value if using IVNT transformations}

\item{age_of_onset_phenotypes}{data frame of age_of_onset phenotypes or a NULL value}

\item{PheWAS_manifest}{manifest document}

\item{age_of_onset_all}{if True creates all age of onset phenotypes}

\item{all_phenotypes}{all available phenotypes}
}
\value{
Phenotypes as dataframe ready for analysis
}
\description{
Allows main conversion to data frame of phenotypes
}
