% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_phenotyping.R
\name{making_case_lists}
\alias{making_case_lists}
\title{Helper function for per_line edit}
\usage{
making_case_lists(x, y, concept_conversion, operator_conversion)
}
\arguments{
\item{x}{PheWAS_ID}

\item{y}{primary care code list}

\item{concept_conversion}{output of concept_list_conversion}

\item{operator_conversion}{output of operator_list_conversion}
}
\value{
A list combining concept_conversion and operator_conversion files for per_line function
}
\description{
Helper function for per_line edit
}
