% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_phenotyping.R
\name{making_case_names}
\alias{making_case_names}
\title{Helper function for per_line edit}
\usage{
making_case_names(x, y)
}
\arguments{
\item{x}{list of numbers, labelling numbered concepts.}

\item{y}{list of numbers, labelling numbered Boolean values.}
}
\value{
A vector of case names.
}
\description{
Helper function for per_line edit
}
