% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_phenotyping.R
\name{making_control_pop_lists}
\alias{making_control_pop_lists}
\title{Makes control populations for composite phenotypes}
\usage{
making_control_pop_lists(x, control_populations)
}
\arguments{
\item{x}{a vector of the control population names.}

\item{control_populations}{control populations datafarame of Ids.}
}
\value{
control populations in phenotype form.
}
\description{
Makes control populations for composite phenotypes
}
\keyword{internal}
