% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables_graphs.R
\name{making_graphs}
\alias{making_graphs}
\title{Makes graphs from converted RDS files.}
\usage{
making_graphs(a, b, c, d, FDR_figure, max_FDR_graph, graph_save_location)
}
\arguments{
\item{a}{type of analysis data}

\item{b}{Type of analysis}

\item{c}{Graph choice}

\item{d}{sex phenotype labels}

\item{FDR_figure}{value for significance threshold for fdr.}

\item{max_FDR_graph}{value for lowest fdr value, converts 0 into this number.}

\item{graph_save_location}{full file location to save graphs}
}
\value{
Creates and then saves graphs from results files.
}
\description{
Makes graphs from converted RDS files.
}
