% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_association_testing.R
\name{making_GRS_folders}
\alias{making_GRS_folders}
\title{Simple helper function for making folders}
\usage{
making_GRS_folders(x, analysis_folder)
}
\arguments{
\item{x}{Per trait map of files}

\item{analysis_folder}{Location where analysis output is saved to.}
}
\value{
Plink association results.
}
\description{
Simple helper function for making folders
}
\keyword{internal}
