% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_filtering.R
\name{minimum_data_R}
\alias{minimum_data_R}
\title{Searches and extracts named columns from data file}
\usage{
minimum_data_R(
  data_folder,
  data_files,
  r_format,
  data_field_ID,
  data_name_pattern,
  N_cores,
  save_loc,
  exclusions
)
}
\arguments{
\item{data_folder}{Full path of the directory that contains the data files that will be formatted and concatenated.}

\item{data_files}{Comma separated full file paths of the data files that will be formatted and concatenated.}

\item{r_format}{Specific to UK Biobank data. Specify if the input for the data have been downloaded using the R option. If the data have been downloaded using the .csv or .txt options, then no input is required (default).}

\item{data_field_ID}{Full path to the file containing the field_IDs required for Deep-PheWAS if not using default file. Is a plain text file with no header one field-ID per row. Field-ID is a numeric value, example field-ID 54 is UK biobank assesment centre.}

\item{data_name_pattern}{Character string for isolating data files if these are in a directory with other files. Defaults to using all files in the directory specified by the data_folder argument.}

\item{N_cores}{Number of cores requested if parallel computing is desired. Defaults to single core computing.}

\item{save_loc}{Full path to save file location for minimum_tab_data.}

\item{exclusions}{Full path to the file containing individuals to be excluded from the analysis. Defaults behaviour is to retain all individuals.}
}
\value{
A dataframe containing only columns of interest
}
\description{
Searches and extracts named columns from data file
}
