% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_phenotyping.R
\name{operator_list_conversion}
\alias{operator_list_conversion}
\title{Helper function for per_line edit}
\usage{
operator_list_conversion(x, y)
}
\arguments{
\item{x}{A sequence of values representing part of the composite phenotype map.}

\item{y}{the instructions for the composite phenotype}
}
\value{
A list which can be edited to create composite phenotypes.
}
\description{
Helper function for per_line edit
}
