% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables_graphs.R
\name{per_group_function}
\alias{per_group_function}
\title{Produces per-group graphs.}
\usage{
per_group_function(
  x,
  y,
  save_root,
  analysis_name,
  graph_choice,
  FDR_figure,
  max_FDR_graph,
  PheWAS_label_filter,
  max_overlap,
  graph_type,
  label_size
)
}
\arguments{
\item{x}{Data table for graphs}

\item{y}{Group name}

\item{save_root}{full file location for save file minus specific file name}

\item{analysis_name}{name of teh analysis}

\item{graph_choice}{numeric value indicating if all or only some graphs are chosen}

\item{FDR_figure}{value of significance for FDR.}

\item{max_FDR_graph}{value for lowest fdr value, converts 0 into this number.}

\item{PheWAS_label_filter}{list of PheWAS_IDs to label within the all_pheno graph.}

\item{max_overlap}{maximum number of overlaps for labelled phenotypes in the all_pheno graphs}

\item{graph_type}{save format of the graphs any input readable from ggsave is accepted}

\item{label_size}{size of the text for labelled phenotypes.}
}
\value{
saved graphs per group
}
\description{
Produces per-group graphs.
}
