% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_association_testing.R
\name{per_group_per_trait_GRS}
\alias{per_group_per_trait_GRS}
\title{Applies association analysis per grouping variable for GRS data.}
\usage{
per_group_per_trait_GRS(
  x,
  GRS_input,
  non_GRS_data,
  covariates,
  all_phenos,
  PheWAS_manifest,
  analysis_save_name,
  quant_min_cases,
  binary_min_cases,
  covariate_col_names,
  old_results
)
}
\arguments{
\item{x}{Group name}

\item{GRS_input}{Data frame guide to location of GRS data, and phenotype data per trait/grouping combination.}

\item{non_GRS_data}{Data frame of non-GRS data, eid column, with each column representing a genetic variable.}

\item{covariates}{Full file path to the covariate file edited for use in association testing, see user guide.}

\item{all_phenos}{Data frame containing all possible phenotypes including age coloumn covariates.}

\item{PheWAS_manifest}{Data frame of the PheWAS_manifest containing information on phenotypes.}

\item{analysis_save_name}{Core save name for the analysis.}

\item{quant_min_cases}{minimum number of quantitative cases.}

\item{binary_min_cases}{minimum number of binary cases.}

\item{covariate_col_names}{column names of covariates.}

\item{old_results}{existing results as list item.}
}
\value{
GRS results as a saved RDS object.
}
\description{
Applies association analysis per grouping variable for GRS data.
}
\keyword{internal}
