% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_phenotyping.R
\name{per_line_edit}
\alias{per_line_edit}
\title{Parses the instruction from composite phenotype map into commands interpretable in R.}
\usage{
per_line_edit(x, phenotypes)
}
\arguments{
\item{x}{Instruction for creation of composite phenotypes.}

\item{phenotypes}{all available phenotypes generated by the previous functions.}
}
\value{
A data frame per PheWAS-ID for composite phenotypes.
}
\description{
Parses the instruction from composite phenotype map into commands interpretable in R.
}
