% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{per_pheno_ancestry}
\alias{per_pheno_ancestry}
\title{Splits phenotype by grouping variable}
\usage{
per_pheno_ancestry(x, y)
}
\arguments{
\item{x}{phenotype dataframe}

\item{y}{Data frame of IDs per group identifier}
}
\value{
phenotype split by population
}
\description{
Splits phenotype by grouping variable
}
\keyword{internal}
