% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phecode_generating.R
\name{phecoding}
\alias{phecoding}
\title{Creates phecode phenotypes for available ICD9 and ICD10 data}
\usage{
phecoding(a, b, c, d, e, f, g, h, i, j, k, l)
}
\arguments{
\item{a}{PheWAS_ID}

\item{b}{phecode}

\item{c}{1st lower phecode range}

\item{d}{1st upper phecode range}

\item{e}{2nd lower phecode range}

\item{f}{2nd upper phecode range}

\item{g}{3rd lower phecode range}

\item{h}{3rd upper phecode range}

\item{i}{if sex specific or not}

\item{j}{data frame of mapped phecodes to ICD10/9 data}

\item{k}{data frame for control exclusions}

\item{l}{data frame fro combined sex}
}
\value{
A dataframe containing phecode status for each participant
}
\description{
Creates phecode phenotypes for available ICD9 and ICD10 data
}
\keyword{internal}
