% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{phenotype_preparation}
\alias{phenotype_preparation}
\title{Filters and prepares phenotyeps for association analysis.}
\usage{
phenotype_preparation(
  phenotype_filtered_save_name,
  phenotype_folder,
  phenotype_files,
  groupings,
  quantitative_Case_N,
  binary_Case_N,
  relate_remove,
  kinship_file,
  sex_split_phenotypes,
  sex_info,
  male,
  female,
  PheWAS_manifest_overide,
  age_of_onset_phenotypes,
  DOB_file,
  IVNT,
  save_RDS,
  stats_save
)
}
\arguments{
\item{phenotype_filtered_save_name}{Full file path for save location of filtered phenotypes. Saves and dataframe per-group (if provided) that is the input to the association testing. Name of final save is /path/to/file/(group)_savename.}

\item{phenotype_folder}{Full path of the folder containing the phenotype data created in previous steps.}

\item{phenotype_files}{Comma separated full file paths of phenotype data created in previous steps.}

\item{groupings}{Full path of the file containing group information used for stratifying the sample. If this argument is specified, the phenotype filtering will be performed within the groups. Otherwise, the filtering will be performed in all individuals.}

\item{quantitative_Case_N}{Number that represents the minimum number of cases for quantitative phenotype inclusion default=100.}

\item{binary_Case_N}{Number that represents the minimum number of cases for binary phenotype inclusion default=50.}

\item{relate_remove}{Specify whether related individuals are additionally excluded (TRUE) or not (FALSE). Default is FALSE.}

\item{kinship_file}{Full path to the file containing kinship coefficient estimates.}

\item{sex_split_phenotypes}{Full file path to file containing single column of PheWAS_IDs to make sex specific phenotypes, will derive and add a male and female version of each phenotype. All sex specific phenotypes be labelled (PheWAS_ID)_male or (PheWAS_ID)_female.}

\item{sex_info}{Full path of the combined_sex file (generated by the 02_data_preparation.R script).}

\item{male}{Value corresponding to males in the combined_sex file must be a number. default=1.}

\item{female}{Value corresponding to females in the combined_sex file must be a number. default=0.}

\item{PheWAS_manifest_overide}{Full file path of the alternative PheWAS_manifest file.}

\item{age_of_onset_phenotypes}{Full file path to file containing three columns PheWAS_ID,lower_limit,upper_limit. PheWAS_ID are the phenotypes to create the age_of_onset phenotypes for, lower_limit is the lower age boundary to filter read as <=, upper_limit is the upper age boundary acceptable read as >=. All age of onset phenotypes to be labelled (PheWAS_ID)_age_of_onset.}

\item{DOB_file}{Full file path to file containing DOB information.}

\item{IVNT}{Specify whether quantitative phenotypes, defined in the PheWAS manifest, should undergo a rank-based inverse normal transformation (TRUE) or not (FALSE). Default is FALSE.}

\item{save_RDS}{Full file path to save location to save and intermediate RDS file that has been filtered by case_N and relatedness (if chosen) but not converted to dataframe for analysis.saves an intermediate RDS file that has been filtered by case_N and relatedness (if chosen) but not converted to dataframe for analysis.}

\item{stats_save}{Full file path to save file for stats, saves summary stats for phenotypes for number of cases and controls in each grouping variable.File name will be appended with _N_filtered.csv or _relate_remove.csv, depending on if relate remove was selected.}
}
\value{
A data frame per grouping variable of prepared and filtered phenotypes used for association analysis.
}
\description{
Filters and prepares phenotyeps for association analysis.
}
