% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_association_testing.R
\name{plink_association_testing}
\alias{plink_association_testing}
\title{Performs association testing using PLINK2.}
\usage{
plink_association_testing(
  analysis_folder,
  phenotype_files,
  covariate,
  variants_for_association,
  analysis_name,
  group_name_overide,
  PheWAS_manifest_overide,
  plink_exe,
  save_plink_tables,
  phenotype_inclusion_file,
  phenotype_exclusion_file,
  split_group,
  N_quant_split,
  N_binary_split,
  split_analysis,
  model,
  check_existing_results
)
}
\arguments{
\item{analysis_folder}{Full path of the directory in which the association results will be saved.}

\item{phenotype_files}{Comma-separated list containing the full paths of the phenotype data.}

\item{covariate}{Full path of a file containing covariate data to be used in model adjustment. See user guide for format.}

\item{variants_for_association}{Full path of the genetic data for the SNPs of interest (produced by the extract_SNP.R script).}

\item{analysis_name}{Name for the analysis to be used in the naming of the output files.}

\item{group_name_overide}{Comma-separated list containing alternative group names. By default, group names are extracted from the suffix of the file names provided in the phenotype_files argument. For example, if a file named /home/phenotypes/EUR_phenotypes.csv were provided, the corresponding group name would be "EUR". This argument allows for a different group name to be specified. The order of names provided should match the files specified in the phenotype_files argument.}

\item{PheWAS_manifest_overide}{Full file path of the alternative PheWAS_manifest file.}

\item{plink_exe}{Command to execute plink2 program.}

\item{save_plink_tables}{Select if wanting to save the raw plink results per group as a table always saved in analysis_folder/assoiation_results/group/group_plink_results_raw.}

\item{phenotype_inclusion_file}{Full file path to a txt file containing single column containing full PheWAS_ID of phenotypes that will be included. Cannot be used with phenotype_exclusion_file argument.}

\item{phenotype_exclusion_file}{Full file path to a txt file containing single column containing full PheWAS_ID of phenotypes that will be excluded. Cannot be used with phenotype_inclusion_file argument.}

\item{split_group}{Comma separated groups that require splitting for more efficient analysis. Does not run the analysis per group but splits and saves the phenotypes into smaller files which can then be analysed using cluster based computing. The number of phenotypes in each split file is dependent on the type (binary or quantitative) and is set using N_quant_split and N_binary_split. Split files are saved in analysis_folder/group_split with group being the group name. A file is created in analysis_folder/group_split saved as group_split_guide with group once again being the group name inputted. This file lists the file names of the splits, which can be used to guide distributed computing analysis.}

\item{N_quant_split}{Number of quantitative phenotypes per split file.}

\item{N_binary_split}{Number of binary phenotypes per split file.}

\item{split_analysis}{Select if the input being analysed is from the split_group command, only changes the saving and attempted combining of files.}

\item{model}{Genetic model to use for analysis, can be one of genotypic, hethom, dominant, recessive, hetonly. If not one of these options will be converted to blank.}

\item{check_existing_results}{Input if wanting to re-run part of the analysis whilst checking for existing results in the plink_results folder.Used primarily where a run has aborted part way through.}
}
\value{
Association test results.
}
\description{
Performs association testing using PLINK2.
}
