% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_split.R, R/plink_association_testing.R
\name{plink_file_edit_join}
\alias{plink_file_edit_join}
\title{Combines PLINK results from the split_groups and split_analysis.}
\usage{
plink_file_edit_join(x, y, remove_string, analysis_name)

plink_file_edit_join(x, y, remove_string, analysis_name)
}
\arguments{
\item{x}{result files}

\item{y}{name of files}

\item{remove_string}{vector to remove}

\item{analysis_name}{name of the analysis}
}
\value{
list of results

Saves data frames
}
\description{
Combines PLINK results from the split_groups and split_analysis.

edits plink results
}
\keyword{internal}
