% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primary_care_quant_phenotypes.R
\name{primarycare_quantitative_phenotypes}
\alias{primarycare_quantitative_phenotypes}
\title{MMakes primary care quantitative phenotypes guided by PheWAS manifest.}
\usage{
primarycare_quantitative_phenotypes(
  GPC,
  DOB,
  phenotype_save_file,
  N_cores,
  PheWAS_manifest_overide
)
}
\arguments{
\item{GPC}{Full path of the primary care clinical data file from UK Biobank.}

\item{DOB}{Full path of a file describing participant date of birth. This does not need to be the exact date. By default for UK Biobank data, the script creates the date of birth from month and year of birth (data-fields 52 and 34).}

\item{phenotype_save_file}{Full path of the save file for the generated concepts RDS to be used for phenotype creation.}

\item{N_cores}{Number of cores requested if parallel computing is desired. Defaults to single core computing.}

\item{PheWAS_manifest_overide}{Full file path of the alternative PheWAS_manifest file.}
}
\value{
An RDS file containing lists of data-frames each representing a single primary-care quantitative phenotype
}
\description{
MMakes primary care quantitative phenotypes guided by PheWAS manifest.
}
