% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables_graphs.R
\name{R_association_function}
\alias{R_association_function}
\title{Produces graphs for R-association data.}
\usage{
R_association_function(
  x,
  save_root,
  analysis_name,
  graph_choice,
  FDR_figure,
  max_FDR_graph
)
}
\arguments{
\item{x}{Data table for graphs}

\item{save_root}{full file location for save file minus specific file name}

\item{analysis_name}{name of teh analysis}

\item{graph_choice}{numeric value indicating if all or only some graphs are chosen}

\item{FDR_figure}{value of significance for FDR.}

\item{max_FDR_graph}{value for lowest fdr value, converts 0 into this number.}
}
\value{
saved graphs per group
}
\description{
Produces graphs for R-association data.
}
