% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_association_testing.R
\name{R_association_testing}
\alias{R_association_testing}
\title{Performs association testing in R for either genetic risk score (GRS) and non-GRS data}
\usage{
R_association_testing(
  analysis_folder,
  phenotype_files,
  covariates,
  GRS_input,
  non_GRS_data,
  group_name_overide,
  PheWAS_manifest_overide,
  analysis_name,
  N_cores,
  phenotype_inclusion_file,
  phenotype_exclusion_file,
  binary_Case_N,
  quantitative_Case_N
)
}
\arguments{
\item{analysis_folder}{Full path to the folder location of the folder hosting this analysis.}

\item{phenotype_files}{Comma separated full file paths of the phenotype data. For example /home/phenotypes/EUR_pheno,/home/phenotypes/AFR_pheno}

\item{covariates}{Full file path to the covariate file edited for use in association testing, see user guide.}

\item{GRS_input}{Full file path of the GRS_input csv file. See user guide for more information on format.}

\item{non_GRS_data}{Full file path to the non_GRS genetic data. See user guide for more information on format.}

\item{group_name_overide}{Comma-separated list containing alternative group names. By default, group names are extracted from the suffix of the file names provided in the phenotype_files argument. For example, if a file named /home/phenotypes/EUR_phenotypes.csv were provided, the corresponding group name would be "EUR". This argument allows for a different group name to be specified. The order of names provided should match the files specified in the phenotype_files argument.}

\item{PheWAS_manifest_overide}{Full file path of the alternative PheWAS_manifest file.}

\item{analysis_name}{Name for the analysis, is used later in saving tables, so should distinguish between other analyses. For GRS analysis this name is always the trait being analysed.}

\item{N_cores}{Number of cores requested if wanting to use parallel computing.}

\item{phenotype_inclusion_file}{Full file path to a txt file containing single column containing full PheWAS_ID of phenotypes that will be included. Cannot be used with phenotype_exclusion_file argument.}

\item{phenotype_exclusion_file}{Full file path to a txt file containing single column containing full PheWAS_ID of phenotypes that will be excluded. Cannot be used with phenotype_inclusion_file argument.}

\item{binary_Case_N}{Number that represents the minimum number of cases for binary phenotype inclusion. Default=50}

\item{quantitative_Case_N}{Number that represents the minimum number of cases for quantitative phenotype inclusion. Default=100}
}
\value{
Association test results as a RDS object.
}
\description{
Performs association testing in R for either genetic risk score (GRS) and non-GRS data
}
