% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_SNPs.R
\name{recoding_variants}
\alias{recoding_variants}
\title{Recodes SNP list input}
\usage{
recoding_variants(x)
}
\arguments{
\item{x}{SNP list input}
}
\value{
Recoded SNPs creates unique value unlike rsid which has duplicates.
}
\description{
Recodes SNP list input
}
