% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{reducing_phenotypes}
\alias{reducing_phenotypes}
\title{Removes phenotypes based on case number.}
\usage{
reducing_phenotypes(x, y)
}
\arguments{
\item{x}{all phenotypes as list.}

\item{y}{PheWAS_IDs to include.}
}
\value{
Phenotypes filtered for case number
}
\description{
Removes phenotypes based on case number.
}
\keyword{internal}
