% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{relate_remove_fn}
\alias{relate_remove_fn}
\title{Removes related individuals from phenotypes.}
\usage{
relate_remove_fn(x, y, call_rate_kinships, PheWAS_manifest)
}
\arguments{
\item{x}{Phenotypes extracted}

\item{y}{PheWAS_ID}

\item{call_rate_kinships}{call_rate_kinship file}

\item{PheWAS_manifest}{manifest file}
}
\value{
A list phenotypes filtered for relatedness.
}
\description{
Removes related individuals from phenotypes.
}
