% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{saving_files}
\alias{saving_files}
\title{Saves list objects as dataframes per grouping variable}
\usage{
saving_files(a, b, save_name_core, save_folder)
}
\arguments{
\item{a}{Phenotypes as list object}

\item{b}{grouping name variable}

\item{save_name_core}{save_name given as argument}

\item{save_folder}{folder for the save_name given as an argument.}
}
\value{
A saved table of phenotypes per grouping name variable.
}
\description{
Saves list objects as dataframes per grouping variable
}
\keyword{internal}
