% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_association_testing.R
\name{saving_files_simple}
\alias{saving_files_simple}
\title{Simple saving function}
\usage{
saving_files_simple(x, y, z, analysis_folder)
}
\arguments{
\item{x}{Df to save}

\item{y}{save name}

\item{z}{Group name.}

\item{analysis_folder}{main folder hosting analsysis}
}
\value{
Saves data frames
}
\description{
Simple saving function
}
\keyword{internal}
