% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{selecting_phenotype_cases_N}
\alias{selecting_phenotype_cases_N}
\title{Applies sum_stats per grouping population for each phenotype returning PheWAS_IDs filtered by case number}
\usage{
selecting_phenotype_cases_N(
  x,
  quant_min_cases,
  binary_min_cases,
  PheWAS_manifest
)
}
\arguments{
\item{x}{phenotype dataframe having been split by grouping variable.}

\item{quant_min_cases}{minimum case number required for quantitative traits.}

\item{binary_min_cases}{minimum case number required for binary traits.}

\item{PheWAS_manifest}{PheWAS_manifest.}
}
\value{
PheWAS_ID meet filtering conditions for case number.
}
\description{
Applies sum_stats per grouping population for each phenotype returning PheWAS_IDs filtered by case number
}
\keyword{internal}
