% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{setdiff_all}
\alias{setdiff_all}
\title{Creates a symmetrical version of setdiff}
\usage{
setdiff_all(vec1, vec2)
}
\arguments{
\item{vec1}{vector 1}

\item{vec2}{vector 2}
}
\value{
a vector of differences
}
\description{
Creates a symmetrical version of setdiff
}
\author{
DIZtools
}
\keyword{internal}
