% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{sex_specific_female}
\alias{sex_specific_female}
\title{Creates female specific versions of existing phenotypes}
\usage{
sex_specific_female(x, y, female_ID)
}
\arguments{
\item{x}{Name of sex-specific phenotype}

\item{y}{Non-sex specific phenotype to be edited}

\item{female_ID}{vector of female IDs}
}
\value{
Female sex-specific phenotype
}
\description{
Creates female specific versions of existing phenotypes
}
\keyword{internal}
