% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{sex_specific_male}
\alias{sex_specific_male}
\title{Creates male specific versions of existing phenotypes}
\usage{
sex_specific_male(x, y, male_ID)
}
\arguments{
\item{x}{Name of sex-specific phenotype}

\item{y}{Non-sex specific phenotype to be edited}

\item{male_ID}{vector of male IDs}
}
\value{
Male sex-specific phenotype
}
\description{
Creates male specific versions of existing phenotypes
}
