% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_SNPs.R
\name{snp_extractor}
\alias{snp_extractor}
\title{Extracts SNPs from bgen or plink binary files into single file that is used for association.}
\usage{
snp_extractor(
  genetic_file_guide,
  SNP_list,
  analysis_folder,
  bgen_input,
  plink_input,
  plink_exe,
  plink_type,
  ref_bgen,
  bgenix_exe,
  variant_save_name,
  no_delete_temp
)
}
\arguments{
\item{genetic_file_guide}{Full path of the completed genetic_file_guide_template.csv.}

\item{SNP_list}{Full path of the file describing which SNPs are to be extracted from the genetic data files ahead of association testing.}

\item{analysis_folder}{Full path of the folder that will contain the data for the SNPs given by SNP_list. A temporary folder, named "temp_plink" by default, will be created within the folder specified by this argument as a place to hold temporary files needed for the SNP extraction process. Unless otherwise requested by specifying the no_delete_temp argument, the temporary folder and its contents will be deleted following successful SNP extraction.}

\item{bgen_input}{Specify that the genetic data files are in .bgen format.}

\item{plink_input}{Specify that the genetic data files are in PLINK format (.bed or .pgen).}

\item{plink_exe}{Full path to the PLINK2 executable.}

\item{plink_type}{Specify whether the PLINK-formatted genetic data are in .bed or .pgen format.}

\item{ref_bgen}{One of the following three values that specifies which allele is to be used as the reference: ref-first (first allele is the reference, default), ref-last (last allele is the reference), red-unknown (last allele is provisionally treated as the reference).}

\item{bgenix_exe}{Full path to the bgenix executable.}

\item{variant_save_name}{Name of the output genetic data files.}

\item{no_delete_temp}{Specify whether the temporary folder containing intermediate files should be retained (TRUE) or deleted (FALSE). Default is TRUE.}
}
\value{
A pgen file of the selected SNPs
}
\description{
Extracts SNPs from bgen or plink binary files into single file that is used for association.
}
