% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_SNPs.R
\name{snp_selection}
\alias{snp_selection}
\title{Extracts SNPs from bgen or plink binary files into single file that is used for association.}
\usage{
snp_selection(
  x,
  analysis_folder,
  snp_guide,
  plink_exe,
  bgenix_exe,
  plink_type,
  snp_list_file,
  plink_input,
  bgen_input,
  ref_bgen
)
}
\arguments{
\item{x}{Chromosome to search}

\item{analysis_folder}{Location of analysis}

\item{snp_guide}{edited version of snp_list}

\item{plink_exe}{command to activate plink2 in system}

\item{bgenix_exe}{command to activate bgenix in system}

\item{plink_type}{bed or pgen file if using plink files}

\item{snp_list_file}{location of snp_list in temp_plink file}

\item{plink_input}{logical value if using plink as store of genetic data}

\item{bgen_input}{logical value if using bgen as store of genetic data}

\item{ref_bgen}{option required for extracting data from begn files.}
}
\value{
A pgen file of the selected SNPs
}
\description{
Extracts SNPs from bgen or plink binary files into single file that is used for association.
}
\keyword{internal}
