% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{spliting_by_population}
\alias{spliting_by_population}
\title{Splits phenotype by grouping variable}
\usage{
spliting_by_population(x, all_phenotypes)
}
\arguments{
\item{x}{group identifier IDs}

\item{all_phenotypes}{all phenotypes loaded in}
}
\value{
all phenotypes split by each population
}
\description{
Splits phenotype by grouping variable
}
