% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{SR_data_out}
\alias{SR_data_out}
\title{Searches and extracts data fields that have multiple records and adds a date for the data collection which is in separate data-field, current version works only with very specific UK Biobank data}
\usage{
SR_data_out(a, b, c, d, e, f)
}
\arguments{
\item{a}{list of the number of times data was collected}

\item{b}{type of data, in the future to be replaced by imputed field-ids}

\item{c}{dataframe containing date of code}

\item{d}{dataframe containing code codes}

\item{e}{name of data source}

\item{f}{dataframe containing date of assessment centre}
}
\value{
A dataframe containing of extracted data, each data item is a row with participant ID, code and source
}
\description{
Searches and extracts data fields that have multiple records and adds a date for the data collection which is in separate data-field, current version works only with very specific UK Biobank data
}
\keyword{internal}
