% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{sum_stats}
\alias{sum_stats}
\title{Summarises the number of cases and controls per_phenotype inputed.}
\usage{
sum_stats(x, y, z, PheWAS_manifest)
}
\arguments{
\item{x}{Phenotype}

\item{y}{group}

\item{z}{PheWAS_ID.}

\item{PheWAS_manifest}{manifest file}
}
\value{
A dataframe of summary stats for phenotypes.
}
\description{
Summarises the number of cases and controls per_phenotype inputed.
}
