% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotype_preparation.R
\name{summary_all_pops}
\alias{summary_all_pops}
\title{Applies sum_stats per grouping population for each phenotype}
\usage{
summary_all_pops(x, y, PheWAS_manifest)
}
\arguments{
\item{x}{phenotype dataframe having been split by grouping variable.}

\item{y}{Grouping variable names.}

\item{PheWAS_manifest}{manifest file}
}
\value{
summary stats per phenotype per grouping variable.
}
\description{
Applies sum_stats per grouping population for each phenotype
}
\keyword{internal}
