% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tables_graphs.R
\name{making_graphs}
\alias{making_graphs}
\title{Makes graphs from converted RDS files.}
\usage{
making_graphs(
  a,
  b,
  c,
  d,
  FDR_figure,
  max_FDR_graph,
  graph_save_location,
  PheWAS_label_filter,
  max_overlap,
  graph_type,
  label_size,
  order_groups_alphabetically,
  order_phenotypes_alphabetically,
  save_all_graphs
)
}
\arguments{
\item{a}{type of analysis data}

\item{b}{Type of analysis}

\item{c}{Graph choice}

\item{d}{sex phenotype labels}

\item{FDR_figure}{value for significance threshold for fdr.}

\item{max_FDR_graph}{value for lowest fdr value, converts 0 into this number.}

\item{graph_save_location}{full file location to save graphs}

\item{PheWAS_label_filter}{list of PheWAS_IDs to label within the all_pheno graph}

\item{max_overlap}{maximum number of overlaps for labelled phenotypes in the all_pheno graphs}

\item{graph_type}{save format of the graphs any input readable from ggsave is accepted}

\item{label_size}{size of the text for labelled phenotypes.}

\item{order_groups_alphabetically}{T or F to order groups on graph alphabetically rather than default lowest FDR.}

\item{order_phenotypes_alphabetically}{T or F to order phenotypes on graphs alphabetically rather than default lowest FDR.}

\item{save_all_graphs}{T or F to always save every graph with or without a significant result (TRUE) or to only save when at least one association is significant (FALSE).}
}
\value{
Creates and then saves graphs from results files.
}
\description{
Makes graphs from converted RDS files.
}
\keyword{internal}
